refactor: convert guildchristmas aaaa
