+++ src/utils/models/Client.ts
@@ -5,10 +5,9 @@ import { doChatReactions } from "../scheduled/clusterjobs/chatreaction";
 import { runModerationChecks } from "../scheduled/clusterjobs/moderationchecks";
 import { updateCache } from "../imghandler";
 import { runLotteryInterval } from "../scheduled/clusterjobs/lottery";
-import { runChristmas } from "../scheduled/clusterjobs/guildchristmas";
 import { runPremiumChecks } from "../scheduled/clusterjobs/premiumexpire";
 import { runEconomySetup } from "../economy/utils";
-import { runCountdowns, updateCounters } from "../guilds/utils";
+import { runChristmas, runCountdowns, updateCounters } from "../guilds/utils";
 
 export class NypsiClient extends Client {
     public cluster: Cluster.Client;
@@ -29,12 +28,12 @@ export class NypsiClient extends Client {
         runPremiumChecks();
         updateCounters(this);
         runCountdowns(this);
+        runChristmas(this);
 
         if (!this.shard.ids.includes(0)) return;
 
         runLotteryInterval(this);
 
-        runChristmas(this);
         runModerationChecks(this);
         doChatReactions(this);
     }
